requirements.txt
ptable
PyYAML
requests
python-dotenv
